add implementation for migrating existing rewards to accumulated rewards
